lpost_tk/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile
#
aiobotocore==2.3.4
    # via -r requirements.in
aiohttp==3.8.1
    # via aiobotocore
aioitertools==0.10.0
    # via aiobotocore
aiomysql==0.1.1
    # via -r requirements.in
aiosignal==1.2.0
    # via aiohttp
anyio==3.6.1
    # via
    #   httpcore
    #   starlette
    #   watchfiles
async-timeout==4.0.2
    # via aiohttp
attrs==22.1.0
    # via
    #   aiohttp
    #   pytest
botocore==1.24.21
    # via aiobotocore
certifi==2022.6.15
    # via
    #   httpcore
    #   httpx
    #   sentry-sdk
charset-normalizer==2.1.0
    # via aiohttp
click==8.1.3
    # via uvicorn
dnspython==2.2.1
    # via -r requirements.in
fastapi==0.79.0
    # via
    #   -r requirements.in
    #   fastapi-utils
fastapi-utils==0.2.1
    # via -r requirements.in
fluent-logger==0.10.0
    # via -r requirements.in
frozenlist==1.3.1
    # via
    #   aiohttp
    #   aiosignal
h11==0.12.0
    # via
    #   httpcore
    #   uvicorn
httpcore==0.15.0
    # via httpx
httptools==0.4.0
    # via uvicorn
httpx==0.23.0
    # via -r requirements.in
idna==3.3
    # via
    #   anyio
    #   rfc3986
    #   yarl
iniconfig==1.1.1
    # via pytest
jmespath==1.0.1
    # via botocore
motor==3.0.0
    # via -r requirements.in
msgpack==1.0.4
    # via fluent-logger
multidict==6.0.2
    # via
    #   aiohttp
    #   yarl
packaging==21.3
    # via pytest
pluggy==1.0.0
    # via pytest
py==1.11.0
    # via pytest
pydantic==1.9.2
    # via
    #   fastapi
    #   fastapi-utils
pymongo==4.2.0
    # via motor
pymysql==1.0.2
    # via aiomysql
pyparsing==3.0.9
    # via packaging
pytest==7.1.2
    # via
    #   -r requirements.in
    #   pytest-asyncio
pytest-asyncio==0.19.0
    # via -r requirements.in
python-dateutil==2.8.2
    # via botocore
python-dotenv==0.20.0
    # via uvicorn
python-snappy==0.6.1
    # via -r requirements.in
pyyaml==6.0
    # via
    #   -r requirements.in
    #   uvicorn
rfc3986[idna2008]==1.5.0
    # via httpx
sentry-sdk==1.9.4
    # via -r requirements.in
six==1.16.0
    # via python-dateutil
sniffio==1.2.0
    # via
    #   anyio
    #   httpcore
    #   httpx
sqlalchemy==1.4.40
    # via fastapi-utils
starlette==0.19.1
    # via fastapi
tomli==2.0.1
    # via pytest
typing-extensions==4.3.0
    # via pydantic
urllib3==1.26.11
    # via
    #   botocore
    #   sentry-sdk
uvicorn[standard]==0.18.2
    # via -r requirements.in
uvloop==0.16.0
    # via uvicorn
watchfiles==0.16.1
    # via uvicorn
websockets==10.3
    # via uvicorn
wrapt==1.14.1
    # via aiobotocore
yarl==1.8.1
    # via aiohttp
zstandard==0.18.0
    # via -r requirements.in
